pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "operator-display-schema"
version = "0.1.0"
description = "A2UI component catalog and wire protocol schema for Operator display rendering"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.hatch.build.targets.wheel]
packages = ["src/operator_display_schema"]

[tool.hatch.build.targets.wheel.force-include]
"catalog" = "operator_display_schema/catalog"
"schema" = "operator_display_schema/schema"
